Fix merge segment issue
