+++ packages/roosterjs-content-model-api/test/publicApi/segment/setFontNameTest.ts
@@ -237,15 +237,7 @@ describe('setFontName', () => {
                         segments: [
                             {
                                 segmentType: 'Text',
-                                text: 'test',
-                                format: {
-                                    fontFamily: 'Arial',
-                                },
-                                isSelected: true,
-                            },
-                            {
-                                segmentType: 'Text',
-                                text: 'test',
+                                text: 'testtest',
                                 format: {
                                     fontFamily: 'Arial',
                                 },
