+++ packages/roosterjs-content-model-api/test/publicApi/segment/setFontSizeTest.ts
@@ -235,15 +235,7 @@ describe('setFontSize', () => {
                         segments: [
                             {
                                 segmentType: 'Text',
-                                text: 'test',
-                                format: {
-                                    fontSize: '10pt',
-                                },
-                                isSelected: true,
-                            },
-                            {
-                                segmentType: 'Text',
-                                text: 'test',
+                                text: 'testtest',
                                 format: {
                                     fontSize: '10pt',
                                 },
