+++ packages/roosterjs-content-model-api/test/publicApi/segment/toggleBoldTest.ts
@@ -229,15 +229,7 @@ describe('toggleBold', () => {
                         segments: [
                             {
                                 segmentType: 'Text',
-                                text: 'test',
-                                format: {
-                                    fontWeight: 'bold',
-                                },
-                                isSelected: true,
-                            },
-                            {
-                                segmentType: 'Text',
-                                text: 'test',
+                                text: 'testtest',
                                 format: {
                                     fontWeight: 'bold',
                                 },
