+++ packages/roosterjs-content-model-api/test/publicApi/segment/toggleItalicTest.ts
@@ -229,15 +229,7 @@ describe('toggleItalic', () => {
                         segments: [
                             {
                                 segmentType: 'Text',
-                                text: 'test',
-                                format: {
-                                    italic: true,
-                                },
-                                isSelected: true,
-                            },
-                            {
-                                segmentType: 'Text',
-                                text: 'test',
+                                text: 'testtest',
                                 format: {
                                     italic: true,
                                 },
