+++ packages/roosterjs-content-model-api/test/publicApi/segment/toggleSubscriptTest.ts
@@ -229,15 +229,7 @@ describe('toggleSubscript', () => {
                         segments: [
                             {
                                 segmentType: 'Text',
-                                text: 'test',
-                                format: {
-                                    superOrSubScriptSequence: 'sub',
-                                },
-                                isSelected: true,
-                            },
-                            {
-                                segmentType: 'Text',
-                                text: 'test',
+                                text: 'testtest',
                                 format: {
                                     superOrSubScriptSequence: 'sub',
                                 },
