+++ packages/roosterjs-content-model-dom/test/modelApi/common/normalizeContentModelTest.ts
@@ -59,7 +59,12 @@ describe('normalizeContentModel', () => {
                         {
                             segmentType: 'Text',
                             format: {},
-                            text: 'test1test2',
+                            text: 'test1',
+                        },
+                        {
+                            segmentType: 'Text',
+                            format: {},
+                            text: 'test2',
                         },
                     ],
                 },
